+++ src/diff/from-tokens.ts
@@ -1,5 +1,0 @@
-import type { TokenDiff } from '../algorithm/myers';
-import type { Diff } from '../types';
-
-export const joinTokenDiffs = (diffs: readonly TokenDiff<string>[]): Diff[] =>
-  diffs.map(([operation, tokens]) => [operation, tokens.join('')]);